+++ bfabric_app_runner/docs/changelog.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Added
 
+- `static_yaml` input type to write parameters etc.
 - `dispatch.dispatch_resource_flow` implements a generic solution for dispatching resource flow workunits without
     having to perform entity look up in many cases yourself.
 - App Definition now supports omitting the collect step.